Hide service config in haproxy-config output
The *_services variables can have a lot of information in them. Besides
cluttering the output, it's possible that over time something sensitive
could make its way into this variable, which would expose that
information over the output log. This changes the log output to only
include the name of the service that is currently being processed.

Change-Id: I4f569f3d5ef19817f24458070eb346a67c9da591